refactor: 将布局与设置组件迁移至对应模块目录
Titlebar、Sider、SidebarNavItem 散落在 components/ 顶层，SettingsSider 位于 pages/settings/ 根目录，目录职责不清晰。将其集中到 layouts/components/ 和 pages/settings/components/ 下，使每个模块的组件归属更明确，便于后续维护和扩展。
